Create a pages directory containing page component, Creat default exports, Create a router file
