+++ src/pages/Account/index.js
@@ -1,0 +1,3 @@
+import Account from './Account';
+
+export default Account;
