+++ src/pages/Category/index.js
@@ -1,0 +1,3 @@
+import Category from './Category';
+
+export default Category;
