+++ src/pages/Logout/index.js
@@ -1,0 +1,3 @@
+import Logout from './Logout';
+
+export default Logout;
