+++ src/pages/index.js
@@ -1,0 +1,6 @@
+import Account from './Account';
+import Category from './Category';
+import Dashboard from './Dashboard';
+import Logout from './Logout';
+
+export { Account, Category, Dashboard, Logout };
