+++ src/router.js
@@ -1,0 +1,6 @@
+export default {
+    Dashboard: '/',
+    Account: '/account',
+    Category: '/category/:department',
+    Logout: '/logout',
+};
